requirement.txt
argparse==1.4.0
pillow==8.0.0
pytorchcv==0.0.58
torchvision==0.7.0
torch==1.6.0
numpy==1.19.2
opencv-python==4.4.0.42
opencv-contrib-python==4.2.0.34